refactor(core,bootstrap,adapters): V2+V3+V-extra complete — #760
V2 outbound_router.py: 324→298L (+outbound_audio.py)
V2 hub_standalone.py:  362→290L (+auth_seeding.py, +hub_builder.py)
V2 cli_pool.py:        406→266L (+cli_pool_streaming.py, +cli_pool_lifecycle.py)
V3 hub.py:             542→298L (+hub_shutdown, hub_circuit_breaker,
                                   hub_pool_delegation, hub_dispatch mixins)
V-extra _shared_streaming.py: 478→31L shim (+_shared_streaming_state.py 195L,
                                              +_shared_streaming_emitter.py 297L)

Layering fix: turn_store_session.py infrastructure→core/stores (V1 leftover);
dropped aiosqlite annotation to keep core/stores protocol-only per ADR-048.

All targeted files now ≤300L. Public APIs preserved.

Gates: ruff ✓ | pyright 0 errors | 2787 tests pass | coverage 83%
