+++ src/lyra/core/hub/__init__.py
@@ -15,6 +15,7 @@ from .middleware import (
     build_default_pipeline as build_default_pipeline,
 )
 from .outbound_dispatcher import OutboundDispatcher as OutboundDispatcher
+from .outbound_router import AudioDispatch as AudioDispatch
 from .outbound_router import OutboundRouter as OutboundRouter
 from .outbound_router import TtsDispatch as TtsDispatch
 from .pipeline_types import (
@@ -26,6 +27,7 @@ from .pipeline_types import (
 
 __all__ = [
     "Action",
+    "AudioDispatch",
     "ChannelAdapter",
     "Hub",
     "IdentityResolver",
